[WIP]Start generalized technology description
First version of python classes for generalized description of micro-
electronics processes. Consists of following main elements:
* 'Property': object to describe properties of process elements
* 'Condition': object to describe rules for the process
* 'Layer': the layers of the process to build up a full stack
* 'Device': object to describe devices of the process
* 'Technlogy': class to bring together all elements that make up a
process and describe the rules that need to be followed.
* 'FreePDK45': example technogy description based on FreePDK45

This is code under heavy development with likely full refactoring and
reimplementation of code happening in near future. It has only been
tested as far as the import of the freepdk45 works in techno_test
notebook. Documentation will be added when API is more stable.