[WIP]Rename Layer and Device
Rename Layer to Mask as representing masks during wafer processing.
Device to Pritimitive as to represent the primitives to be used when
designing circuit for a technology.
Current rationale is that everything in a design has to be represented
by a primitive and mask are generated out of that.
Primitive are also the main object that determines the rules for making
a design.