Made separate DerivedWire primitive
Wire primitive is now based on a Mask object, DerivedWire on a Wire object with
extra marker layers. Used DerivedWire to implement the ntap and ptap for
FreePDK45.
Don't allow grid parameter for DerivedWire, it only makes sense for Wire.
Allow DerivedWire for top and bottom of Via and active and poly for MOSFERGate.