Introduce _MaskPrimitive
The _MaskPrimitive class is a primitive based refering to a mask.Mask object.
For rule generation a mask need to be available so this class is there to
ensure that. _WidthSpacePrimitive, Implant, Deposition, Marker,
Via are now _MaskPrimitives.
Define a grid property for _MaskPrimitive that allows to override technology
grid for a specific mask.
Introduced Wafer class in mask.py to be for a wafer Marker primitive. Removed
Substrate class as it will now also be presented with a Marker primitive.