mask.py: add Connect and SameNet
class Connect is a maks._Mask and rule._Rule subclass. The rule allows to
indicate that some masks are connected to each other; being a mask allows
to use some a mask property to create conditions to be used as rules.
These classes are now used to generate the rules for min_space_samenet
parameter of primitive.Well. Also an exception will now be raised when
that parameter is specified but the well is not connected to a DerivedWire.