technology.Technology: replace _checks() with _build_interconnect()
Now the interconnect check will done during building of the
interconnection of all the primitives.
The primitives will also be ordered from bottom to top so that primitives
which are used in another primitive come before in order before that
primitive.
In the freepdk45 noteboook the rules are not sorted anymore as they will
not be sorted due to the sorting of the primitives.