Added PrimitiveDispatcher class
Added in technology/dispatcher.py. This is a helper class that allows to
execute code specific for a certain primitive type. If no code is added
for a certain type code for parent class will be called.
This is inspired by the visitor pattern but without the need for a accept
method on the objects one wants to visit.