Added pdkmaster/design/circuit.py
This contains the _Circuit class and support classes. These allow to
describe a circuit. A circuit consists of instances of technology
primitives.
A CircuitFactory class is added for circuit object generation, this
allows to have one 'factory' per technology and avoids having to redo
tech adaption for each Circuit separately.