property_.py: Add Enclosure class
This class represent a value for an enclosure of two features. It can
either by a single float representing an enclosure value the same for
all directions and a pair of float representing an enclosure for on
direction and the perpendicular direction.
Also removed default value for min_(bottom|top)_enclosure for Via. It now
needs to be explicitly specified for all Vias.