circuit._InstanceNet: name handling changes
Now name of the instancenet is the name of the net and extra attribute is
added that gives the '_instname_._netname_' name.
Added _InstanceNets TypedTuple now that will index on the full name. The
childports atrribute of Net will use this to allow having net from
different instances that have the same name inside the instance.
The ports attribute of Instance is still indexed on the instance net name.
This will allow the following code to (still) work:

    vdd.childports += (pmos1.ports.bulk, pmos2.ports.bulk)