Add design.library
Introduced the Library class and support classes _Cell. _Cells, _CellLayout
and _CellLayouts.
A library consists of a number of cells. Currently each cell can have
a number of circuits and a number of layouts. The default circuit and
layout are the ones with the same name as the cell.
Currently no consistency check is implemented between the different views.

Extended pdkmaster export to handle the newly introduced Library class.