Add class StdCellLibrary.
Optional global_nets attribute is added to Library. This alloww
to specify nets that are globally connected.
RoutingGauge support class is added. This class allows to specify
the routing gauge(s) for a library.
StdCellLibrary class is added which is a subclass of Library and
as the name implies is to be used for standard cell libraries.
It has extra arguments to __init__() method:
- global_nets has to be specified.
- routtinggauge: RoutingGauge object(s) to define preferred
  routing configurations
- pingrid_pitch: the pitch of the pin grid of the standard cells
- row_height: the height of one row of standard cells.