Add computed properties for Technology objects.
Added facility to compute minimal width, space and pitch for different
layers of a technology. All relevant primitive specs will be taken into
account to compute these values.
The purpose is to extend in the future so the amount of computation that
needs to be done by layout generators can be minimized.