Add support for blockage layers.
An extra blockage parameter is added to the WaferWire, GateWire, MetalWire
and Via primitive. This parameter has to be a Marker primitive. The
interpretation of a blockage layer may be flow dependent but most of the
time it indicates to the P&R tool places where no shapes should be drawn.