library.py: Cell with on-demand circuit and layout generation.
Added _OnDemandCell class that allows to generate a cell that will only
generate circuit and/or layout the first time it is accessed.
This should allow to speed up the import of a technology that generates
libraries on that event. Time is saved because not all the circuits and the
layouts need to be generated for all cells in it's libraries. They can now
be generated when actually used by user code.
Update coriolis export code to be sure all layouts are generated before doing
the actual export.
Also changed layout existence check for a layout by trying to access the layout
attribute. This gives better error messages when there are errors in
on-demand creation of the layout.